fix: headers to prevent buffering SSE response streams
refactor: move 'completions.openai_chat_completion' to 'views.completions'

It is entirely wrapped up in FastAPI.

Closes #682.
