+++ tests/unit/views/test_completions_views.py
@@ -5,6 +5,7 @@ import pytest
 
 from soliplex import installation
 from soliplex import models
+from soliplex import views
 from soliplex.config import completions as config_completions
 from soliplex.views import completions as completions_views
 
@@ -118,6 +119,66 @@ async def test_get_chat_completion(fc, completion_configs):
 
 
 @pytest.mark.anyio
+@pytest.mark.parametrize("w_exception", [False, True])
+@pytest.mark.parametrize("w_history", [False, True])
+@mock.patch("soliplex.completions.stream_chat_responses")
+@mock.patch("fastapi.responses.StreamingResponse")
+async def test_openai_chat_completion(sr_class, scp, w_history, w_exception):
+    CR_MESSAGES = [
+        {
+            "role": "user",
+            "content": "Why is blue?",
+        },
+    ]
+    agent = mock.Mock(spec_set=["model", "run_stream"])
+    agent_deps = mock.Mock(spec_set=())
+
+    if w_history:
+        CR_MESSAGES.insert(
+            0,
+            {
+                "role": "model",
+                "content": "hello",
+            },
+        )
+
+    chat_request = mock.create_autospec(models.ChatCompletionRequest)
+    chat_request.model_dump.return_value = {
+        "model": "NOT USED",
+        "messages": CR_MESSAGES,
+    }
+
+    if w_exception:
+        scp.side_effect = ValueError("testing-stream_chat_responses")
+
+    if w_exception:
+        with pytest.raises(fastapi.HTTPException):
+            await completions_views.openai_chat_completion(
+                agent,
+                agent_deps,
+                chat_request,
+            )
+        return
+    else:
+        response = await completions_views.openai_chat_completion(
+            agent,
+            agent_deps,
+            chat_request,
+        )
+
+    assert response is sr_class.return_value
+
+    sr_class.assert_called_once_with(
+        scp.return_value,
+        media_type="text/event-stream",
+        headers=views.HEADERS_DO_NOT_BUFFER_SSE,
+    )
+
+    # TODO not passing history
+    scp.assert_called_once_with(agent, agent_deps, "Why is blue?", [])
+
+
+@pytest.mark.anyio
 @pytest.mark.parametrize(
     "w_auth_user, exp_user",
     [
@@ -154,7 +215,7 @@ async def test_post_chat_completion_miss(w_auth_user, exp_user):
     ],
 )
 @pytest.mark.parametrize("w_msg", [False, True])
-@mock.patch("soliplex.completions.openai_chat_completion")
+@mock.patch("soliplex.views.completions.openai_chat_completion")
 async def test_post_chat_completion_hit(
     occ,
     w_msg,
